Fix entities
