--- src/entities/book_excerpts_tags.rs
@@ -1,46 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.0
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "book_excerpts_tags")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub book_excerpt_id: Uuid,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub tag_id: Uuid,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::book_excerpt::Entity",
-        from = "Column::BookExcerptId",
-        to = "super::book_excerpt::Column::Id",
-        on_update = "NoAction",
-        on_delete = "Cascade"
-    )]
-    BookExcerpt,
-    #[sea_orm(
-        belongs_to = "super::tag::Entity",
-        from = "Column::TagId",
-        to = "super::tag::Column::Id",
-        on_update = "NoAction",
-        on_delete = "Cascade"
-    )]
-    Tag,
-}
-
-impl Related<super::book_excerpt::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::BookExcerpt.def()
-    }
-}
-
-impl Related<super::tag::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Tag.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
