--- src/entities/mission_memos_tags.rs
@@ -1,46 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.0
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "mission_memos_tags")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub mission_memo_id: Uuid,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub tag_id: Uuid,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::mission_memo::Entity",
-        from = "Column::MissionMemoId",
-        to = "super::mission_memo::Column::Id",
-        on_update = "NoAction",
-        on_delete = "Cascade"
-    )]
-    MissionMemo,
-    #[sea_orm(
-        belongs_to = "super::tag::Entity",
-        from = "Column::TagId",
-        to = "super::tag::Column::Id",
-        on_update = "NoAction",
-        on_delete = "Cascade"
-    )]
-    Tag,
-}
-
-impl Related<super::mission_memo::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::MissionMemo.def()
-    }
-}
-
-impl Related<super::tag::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Tag.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
